Added add functionality in main menu

+++ user_manager.sh
@@ -6,7 +6,7 @@ PASSWORD=""
 # Functions
 generatePassword() {
     local len=$1
-    # echo "Len of pwd: $len"
+    echo "Len of pwd: $len"
     PASSWORD=$(openssl rand -base64 $len)
     echo "PWD: $PASSWORD"
 }
@@ -16,9 +16,16 @@ addUser () {
     then
         echo "User alreay exists"
     else
+        local lenPwd=$2
+
+        if [[ $2 == *[a-zA-z'!'@#\$%^\&*()_+]* || $2 -lt 12 || $2 -gt 24 ]]
+        then
+            lenPwd=16
+        fi
+
         useradd -m -s /bin/bash $1
         echo "User $1 added"
-        generatePassword 16
+        generatePassword $lenPwd
         echo "$1:$PASSWORD" | sudo chpasswd
     fi
 }
@@ -28,7 +35,13 @@ modifyUser () {
 }
 
 deleteUser () {
-    userdel -r -f $1
+    if id "$1" >/dev/null 2>&1;
+    then
+        userdel -r -f $1
+        echo "User $1 deleted"
+    else
+        echo "User doesn't exist"
+    fi
 }
 
 showUsersList() {
@@ -49,30 +62,90 @@ showUsersList() {
     done
 }
 
+showAddUserMenu() {
+
+    local name=""
+    local lenPwd=0
+
+    #-a $lenPwd -eq 0 
+    while [ -z "$name" ] 
+    do
+        read -p "Enter username: " name
+
+        if id "$name" >/dev/null 2>&1 ;
+        then
+            name=""
+            echo "User alreay exists"
+            continue
+        fi
+
+        if [[ $name == [0-9]* ]]
+        then
+            name=""
+            echo "User can't start with a number"
+        fi
+    done
+
+
+    while [ $lenPwd -eq 0 ] 
+    do
+        read -p "Enter password length (min: 12 | max: 24): " lenPwd
+
+        if [[ -z "$lenPwd" || "$lenPwd" == *[a-zA-z'!'@#\$%^\&*()_+]* || $lenPwd -lt 12 || $lenPwd -gt 24 ]]
+        then
+            lenPwd=0
+            echo "Password length is not valid"
+        fi
+    done
+
+    addUser $name $lenPwd
+}
+
+showMainMenu() {
+    local options=("List Users" "Add User" "Delete User" "Modify User [NAME]" "Quit")
+    local width=25
+    local cols=3
+
+    for ((i=0;i<${#options[@]};i++)); do 
+        string="$(($i+1))) ${options[$i]}"
+        printf "%s" "$string"
+        printf "%$(($width-${#string}))s" " "
+        [[ $(((i+1)%$cols)) -eq 0 ]] && echo
+    done
+    echo
+}
+
 # Main Menu Logic
+clear
 if [ $# -eq "0" ]
 then
-    select var in "Add User" "Delete User" "Modify User" "Quit"
-    do
-        echo "$var selected"
+
+    while true; do
+        showMainMenu
+        read -p '#? ' var
+        clear
         case $var in
-            "Add User")
-                echo "Mark selected"
+            1)
+                showUsersList
                 ;;
-            "Delete User")
-                echo "Mark selected"
+            2)
+                showAddUserMenu
                 ;;
-            "Modify User")
-                echo "Mark selected"
+            3)
+                echo "Delete User"
                 ;;
-            "Quit")
+            4)
+                echo "Modify User"
+                ;;
+            5)
                 exit
                 ;;
             *)
-                echo "Named not found"
+                echo "Selected option is not valid"
                 ;;
         esac
     done
+
 else
 
     # echo "Param 2: $2"
@@ -82,7 +155,7 @@ else
             then
                 echo "The username is empty or starts with a number"
             else
-                addUser $2
+                addUser $2 $3
             fi
             ;;
         "delete")
